fixed test error
