+++ io/gff/example_test.go
@@ -1,7 +1,11 @@
 package gff_test
 
 import (
+	"bytes"
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/TimothyStiles/poly/io/gff"
@@ -18,6 +22,48 @@ func Example_basic() {
 		}
 	}
 	// Output: 2.7.1.39
+}
+
+func ExampleRead() {
+	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
+	fmt.Println(sequence.Meta.Name)
+	// Output: U00096.3
+}
+
+func ExampleParse() {
+	file, _ := os.Open("../../data/ecoli-mg1655-short.gff")
+	sequence, _ := gff.Parse(file)
+
+	fmt.Println(sequence.Meta.Name)
+	// Output: U00096.3
+}
+
+func ExampleBuild() {
+	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
+	gffBytes, _ := gff.Build(sequence)
+	gffReader := bytes.NewReader(gffBytes)
+	reparsedSequence, _ := gff.Parse(gffReader)
+
+	fmt.Println(reparsedSequence.Meta.Name)
+	// Output: U00096.3
+}
+
+func ExampleWrite() {
+	tmpDataDir, err := ioutil.TempDir("", "data-*")
+	if err != nil {
+		fmt.Println(err.Error())
+	}
+	defer os.RemoveAll(tmpDataDir)
+
+	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
+
+	tmpGffFilePath := filepath.Join(tmpDataDir, "ecoli-mg1655-short.gff")
+	_ = gff.Write(sequence, tmpGffFilePath)
+
+	testSequence, _ := gff.Read(tmpGffFilePath)
+
+	fmt.Println(testSequence.Meta.Name)
+	// Output: U00096.3
 }
 
 func ExampleGff_AddFeature() {
@@ -41,7 +87,7 @@ func ExampleGff_AddFeature() {
 	_ = sequence.AddFeature(&feature)
 
 	// get the GFP feature sequence string from the sequence struct.
-	featureSequence := feature.GetSequence()
+	featureSequence, _ := feature.GetSequence()
 
 	// check to see if the feature was inserted properly into the sequence.
 	fmt.Println(gfpSequence == featureSequence)
@@ -69,7 +115,7 @@ func ExampleFeature_GetSequence() {
 	_ = sequence.AddFeature(&feature)
 
 	// get the GFP feature sequence string from the sequence struct.
-	featureSequence := feature.GetSequence()
+	featureSequence, _ := feature.GetSequence()
 
 	// check to see if the feature was inserted properly into the sequence.
 	fmt.Println(gfpSequence == featureSequence)
@@ -119,7 +165,7 @@ func TestFeature_GetSequence(t *testing.T) {
 	_ = sequence.AddFeature(&feature)
 
 	// get the GFP feature sequence string from the sequence struct.
-	featureSequence := feature.GetSequence()
+	featureSequence, _ := feature.GetSequence()
 
 	// check to see if the feature was inserted properly into the sequence.
 	if gfpSequence != featureSequence {
