+++ io/gff/gff_test.go
@@ -1,8 +1,6 @@
 package gff_test
 
 import (
-	"bytes"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -57,48 +55,6 @@ func TestGffIO(t *testing.T) {
 		t.Errorf("Build() does not output the same file as was input through ReadGff(). Got this diff:\n%s", gffDiffText)
 	}
 
-}
-
-func ExampleRead() {
-	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
-	fmt.Println(sequence.Meta.Name)
-	// Output: U00096.3
-}
-
-func ExampleParse() {
-	file, _ := os.Open("../../data/ecoli-mg1655-short.gff")
-	sequence, _ := gff.Parse(file)
-
-	fmt.Println(sequence.Meta.Name)
-	// Output: U00096.3
-}
-
-func ExampleBuild() {
-	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
-	gffBytes, _ := gff.Build(sequence)
-	gffReader := bytes.NewReader(gffBytes)
-	reparsedSequence, _ := gff.Parse(gffReader)
-
-	fmt.Println(reparsedSequence.Meta.Name)
-	// Output: U00096.3
-}
-
-func ExampleWrite() {
-	tmpDataDir, err := ioutil.TempDir("", "data-*")
-	if err != nil {
-		fmt.Println(err.Error())
-	}
-	defer os.RemoveAll(tmpDataDir)
-
-	sequence, _ := gff.Read("../../data/ecoli-mg1655-short.gff")
-
-	tmpGffFilePath := filepath.Join(tmpDataDir, "ecoli-mg1655-short.gff")
-	_ = gff.Write(sequence, tmpGffFilePath)
-
-	testSequence, _ := gff.Read(tmpGffFilePath)
-
-	fmt.Println(testSequence.Meta.Name)
-	// Output: U00096.3
 }
 
 func BenchmarkReadGff(b *testing.B) {
